Move store directory
