+++ components/NavigationDrawer.tsx
@@ -5,7 +5,7 @@ import invariant from 'tiny-invariant';
 import classNames from 'classnames/bind';
 import { isOneOf } from '@/types/guard';
 import { useAppDispatch, useAppSelector } from '@/lib/hooks/index';
-import { openSidebar, closeSidebar } from '@/store/todoSlice';
+import { openSidebar, closeSidebar } from '@/lib/store/todoSlice';
 import dayjs from '@/lib/plugins/dayjs';
 import styles from './NavigationDrawer.module.scss';
