+++ components/SearchBox.tsx
@@ -2,7 +2,7 @@ import { useRef, FormEventHandler, FocusEventHandler } from 'react';
 import { useRouter } from 'next/router';
 import classNames from 'classnames/bind';
 import { useAppDispatch, useAppSelector } from '@/lib/hooks/index';
-import { openSearchBox, closeSearchBox } from '@/store/todoSlice';
+import { openSearchBox, closeSearchBox } from '@/lib/store/todoSlice';
 import styles from './SearchBox.module.scss';
 
 const cx = classNames.bind(styles);
