+++ lib/hooks/index.ts
@@ -1,5 +1,5 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import type { RootState, AppDispatch } from '@/store/index';
+import type { RootState, AppDispatch } from '@/lib/store/index';
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
