+++ pages/_app.tsx
@@ -3,7 +3,7 @@ import type { AppProps } from 'next/app';
 import '@fortawesome/fontawesome-free/css/all.css';
 import 'react-datepicker/dist/react-datepicker.css';
 import { Provider } from 'react-redux';
-import { store } from '@/store/index';
+import { store } from '@/lib/store/index';
 import './_app.scss';
 import AppContainer from '@/layouts/AppContainer';
